Phase 1 Task 2 push.
Added Game class which represents all the mechanics of the game, takes user input to make the player do actions, and logs relevant information about the state of the game to the player, allowing them to play the game.

Also modified main to instantiate and run a new game when called.